requirements.txt
AWSIoTPythonSDK==1.2.0
Django==1.11.6
python-interface==1.2.0